+++ src/prime_rl/orchestrator_v2/metrics.py
@@ -10,7 +10,6 @@ Pure-ish — only state is the config. No I/O, no side effects.
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from typing import Any
 
 import pandas as pd
@@ -18,22 +17,7 @@ import verifiers as vf
 
 from prime_rl.configs.orchestrator import OrchestratorConfig
 from prime_rl.orchestrator.vf_utils import get_seq_len
-from prime_rl.orchestrator_v2.ckpt import Progress
-
-
-@dataclass
-class ProcessResult:
-    """Per-batch stats the metrics builder reads. Produced by
-    ``TrainSink.process_batch`` and handed back to the orchestrator (which
-    forwards it to ``MetricsBuilder.build``)."""
-
-    n_trainable: int
-    num_prefill_tokens: int
-    num_decode_tokens: int
-    rollout_prefill_lens: list[int]
-    rollout_decode_lens: list[int]
-    samples_per_rollout: list[int]
-    samples_shipped: int
+from prime_rl.orchestrator_v2.types import ProcessResult, Progress
 
 
 class MetricsBuilder:
